requirements.txt
streamlit
opencv-python
numpy
Pillow
ultralytics==8.0.20
